arch: Add bootloader definitions and rules
The current bootloader definitions and uses cases are both platform and
or board specific. This add bootloader configurations and rules to allow
agnostic and better use cases. It avoid duplicate code and define unique
symbols to standardize implementations.

It add MCUboot defintions, which is current in use, and rework samv7 use
cases.

Signed-off-by: Gerson Fernando Budke <nandojve@gmail.com>